Cut deck, gave an opponent hand, and scaled cards
I added pretty much the same thing as the player hand removing cards from deck and put in array but for opponent. I added a cut the deck for the top card. I added a way to display all cards that should be by using an array i added to my current  display images.  And I spent about an unreasonable amount of time trying to make a loop to look at every possible combination of cards in hands and still don't know how.